added swarm_select ui + optimised swarm algo
