+++ src/main/java/swarm_wars_library/entities/ENTITY.java
@@ -16,5 +16,8 @@ public enum ENTITY {
   PLAYER4_BULLET, 
   TURRET, 
   TURRET_BULLET, 
-  STAR; 
+  STAR,
+  PLAYERUI,
+  PLAYERUI_BOT,
+  PLAYERUI_BULLET,
 }
