+++ src/main/java/swarm_wars_library/entities/Tag.java
@@ -8,15 +8,21 @@ public class Tag {
     if(tag.equals(ENTITY.PLAYER2)){
       return ENTITY.PLAYER2_BULLET;
     }
+    if(tag.equals(ENTITY.PLAYERUI)){
+      return ENTITY.PLAYERUI_BULLET;
+    }
     return ENTITY.TURRET_BULLET;
   }
 
-  public static ENTITY getMothershipTag(ENTITY tag){
+  public static ENTITY getMotherShipTag(ENTITY tag){
     if(tag.equals(ENTITY.PLAYER1_BOT)){
       return ENTITY.PLAYER1;
     }
     if(tag.equals(ENTITY.PLAYER2_BOT)){
       return ENTITY.PLAYER2;
+    }
+    if(tag.equals(ENTITY.PLAYERUI_BOT)){
+      return ENTITY.PLAYERUI;
     }
     return ENTITY.TURRET;
   }
@@ -27,6 +33,9 @@ public class Tag {
     }
     if(tag.equals(ENTITY.PLAYER2_BULLET)){
       return ENTITY.PLAYER2;
+    }
+    if(tag.equals(ENTITY.PLAYERUI_BULLET)){
+      return ENTITY.PLAYERUI;
     }
     return ENTITY.TURRET;
   }
