+++ src/main/java/swarm_wars_library/game_screens/GAMESCREEN.java
@@ -3,6 +3,7 @@ package swarm_wars_library.game_screens;
 public enum GAMESCREEN {
   START,
   FSMUI,
+  SWARMSELECT,
   GAME,
   GAMEOVER
 }
